Build on win32
This is pretty rough code, but it works on my machine.  There are still
several errors in the win32 section of easy_files due to many functions
in the windows api not handling '..' in the file path.  However, most of
these code paths are not currently used.  This at least provides an
example of how to fix those in the future.

The build.bat script will need to be modified with the SDL library and
include directories on your machine.